+++ components/incidents/client-incident-view/about-this-page/BarsDataGenerator.tsx
