+++ components/incidents/client-incident-view/past-incidents-component/PastIncidents.tsx
@@ -5,10 +5,10 @@ import * as React from "react";
 // components
 import { Block } from "baseui/block";
 import { Spinner } from "baseui/spinner";
-import { GetPastIncidentComponents } from "./HelperFunctions";
+import { GetPastIncidentComponents } from "../HelperFunctions";
 
 // constants
-import { NEXT_PUBLIC_AUTH_TOKEN, PAGE_ID } from "../../../constants";
+import { NEXT_PUBLIC_AUTH_TOKEN, PAGE_ID } from "../../../../constants";
 
 /**
  * Fetch components from API
