+++ components/incidents/client-incident-view/styles/scheduledMaintenanceList.tsx
@@ -8,6 +8,7 @@ export const maintenanceListItem: BlockProps = {
         display: "flex",
         flexDirection: "column",
         gap: "10px",
+        margin: "10px 0px",
       },
     },
   },
